spimex/requirements.txt
pandas
sqlalchemy
psycopg2-binary
asyncpg
python-dotenv
requests
aiohttp
alembic
xlrd
tqdm